Bubble up inherited styles.
And here we are, at yet another crux of why CSS is complicated. If all children
of a DOM node share the same inherited attribute (e.g. color or font-size), then
the parent should simply declare the attribute. This is not the same as having
a CSS class applied to all children; this is using the inheriting phase of style
value calculation.